Use NavigationService from InfiniTime, improve nimble headers
Improve the nimble headers such that we can use the NavigationService
from InfiniTime instead of having a slightly modified verison in
InfiniSim.
