--- sim/components/ble/NavigationService.h
@@ -1,63 +1,0 @@
-/*  Copyright (C) 2021  Adam Pigg
-
-    This file is part of InfiniTime.
-
-    InfiniTime is free software: you can redistribute it and/or modify
-    it under the terms of the GNU General Public License as published
-    by the Free Software Foundation, either version 3 of the License, or
-    (at your option) any later version.
-
-    InfiniTime is distributed in the hope that it will be useful,
-    but WITHOUT ANY WARRANTY; without even the implied warranty of
-    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-    GNU General Public License for more details.
-
-    You should have received a copy of the GNU General Public License
-    along with this program.  If not, see <https://www.gnu.org/licenses/>.
-*/
-#pragma once
-
-#include <cstdint>
-#include <string>
-//#define min // workaround: nimble's min/max macros conflict with libstdc++
-//#define max
-//#include <host/ble_gap.h>
-//#include <host/ble_uuid.h>
-//#undef max
-//#undef min
-
-namespace Pinetime {
-  namespace System {
-    class SystemTask;
-  }
-  namespace Controllers {
-
-    class NavigationService {
-    public:
-      explicit NavigationService(Pinetime::System::SystemTask& system);
-
-      void Init();
-
-//      int OnCommand(uint16_t conn_handle, uint16_t attr_handle, struct ble_gatt_access_ctxt* ctxt);
-
-      std::string getFlag();
-
-      std::string getNarrative();
-
-      std::string getManDist();
-
-      int getProgress();
-
-    private:
-//      struct ble_gatt_chr_def characteristicDefinition[5];
-//      struct ble_gatt_svc_def serviceDefinition[2];
-
-      std::string m_flag;
-      std::string m_narrative;
-      std::string m_manDist;
-      int m_progress = 0;
-
-      Pinetime::System::SystemTask& m_system;
-    };
-  }
-}
