+++ sim/host/ble_gatt.cpp
