+++ sim/nrf_assert.h
@@ -1,1 +1,6 @@
+#pragma once
+#include <cassert>
+
+#ifndef ASSERT
 #define ASSERT(expr)   assert(expr)
+#endif
